Improve info lookup for better on hover tooltips
Replace getWordAtPosition with new getSymbolAtPosition

This function is used to extract the string to pass to the ghc-mod info
command. The previous heuristic was to identify a word by finding the
nearest spaces to either side of the given position. Since spacing is often
not required in Haskell, this could fail.

Based on looking at the Haskell report there are two things it is valid
to run info on, identifiers (varid | conid) and operators (varsym | consym)
Report: https://www.haskell.org/onlinereport/lexemes.html

These two have non-overlapping sets of valid characters, basically
letters vs punctuation, so it is possible to tell from looking at the
single character at the given position whether it is an identifier or an
operator and then expand appropriately to the entire symbol by testing
the characters to the left and right to see if they match the production
rule.